Rearranged dashboard components and other screens as well.
