+++ src/App.js
@@ -3,7 +3,6 @@ import { Outlet, Link } from "react-router-dom";
 import Joyride from 'react-joyride';
 import { tourSteps } from './onboarding/tourSteps';
 import './App.css';
-import Calendar from 'react-calendar';
 import PatientList from './components/PatientList';
 import Patient from './components/Patient';
 
@@ -17,40 +16,9 @@ import Messages from './pages/Messages';
 import Schedule from './pages/Schedule';
 
 
-function isSameDay (date1, date2) {
-    return date1.getDate() === date2.getDate() && date1.getMonth() === date2.getMonth() && date1.getFullYear() === date2.getFullYear();
-}
-
-//const datesToAddContentTo = [tomorrow, in3Days, in5Days];
-const datesToAddContentTo = [new Date(2025, 1, 1), new Date(2022, 2, 1), new Date(2022, 3, 1)];
-
-function tileContent({ date, view }) {
-  // Add class to tiles in month view only
-  if (view === 'month') {
-    // Check if a date React-Calendar wants to check is on the list of dates to add class to
-    if (datesToAddContentTo.find(dDate => isSameDay(dDate, date))) {
-      return 'My content';
-    }
-  }
-}
-
-function tileClassName({ date, view }) {
-    const datesToAddClassTo = datesToAddContentTo;
-  // Add class to tiles in month view only
-  if (view === 'month') {
-    // Check if a date React-Calendar wants to check is on the list of dates to add class to
-    if (datesToAddClassTo.find(dDate => isSameDay(dDate, date))) {
-      return 'myClassName';
-    }
-  }
-}
-
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 
 function Home() {
-    const [calendarValue, setCalendarValue] = useState(new Date());
-    function onCalendarChange(nextValue) {setCalendarValue(nextValue);}
-
     const [runTour, setRunTour] = useState(true);
 
     return (
@@ -60,8 +28,6 @@ function Home() {
                 <Topbar />
 
                 <div className="main-content">
-                    <Dashboard />
-                    <Calendar onChange={onCalendarChange} value={calendarValue} tileContent={tileContent} tileClassName={tileClassName} />
                     <main>
                         {/* This is where the nested routes will render */}
                         <Outlet />
@@ -120,10 +86,12 @@ const router = createBrowserRouter([
         path: "/",
         element: <Home />,
         children: [
+            { path: '/', element: <Dashboard /> },
             { path: "about", element: <About /> },
             { path: "contact", element: <Contact /> },
             { path: "patients", element: <PatientList /> },
             { path: "patients/:id", element: <Patient /> },
+            { path: "schedule", element :<Schedule /> }
         ],
         errorElement: <ErrorPage />,
     },
